+++ conduit/__init__.py
@@ -3,7 +3,7 @@ from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
 from flask_jwt_extended import JWTManager
 from flask_marshmallow import Marshmallow
-from app.errors import Errors
+from .errors import Errors
 
 migrate = Migrate()
 jwt = JWTManager()
